:sparkles: feat: Exception handling in products

+++ src/main/java/com/umcosmeticos/umcosmeticos/services/ProductService.java
@@ -3,10 +3,14 @@ package com.umcosmeticos.umcosmeticos.services;
 import com.umcosmeticos.umcosmeticos.dto.ProductDTO;
 import com.umcosmeticos.umcosmeticos.entities.Product;
 import com.umcosmeticos.umcosmeticos.repositorys.ProductRepository;
+import com.umcosmeticos.umcosmeticos.services.exceptions.DatabaseException;
+import com.umcosmeticos.umcosmeticos.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -18,11 +22,9 @@ public class ProductService {
     //Torna a transação mais rápida
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Optional<Product> result = repository.findById(id);
-        Product product = result.get();
-
-        ProductDTO productDTO = new ProductDTO(product);
-        return productDTO;
+        Product product = repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado"));
+        return new ProductDTO(product);
     }
 
     @Transactional(readOnly = true)
@@ -41,15 +43,22 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDTO(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+            Product entity = repository.getReferenceById(id);
+            copyDTO(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Produto não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integração referencial");
+        }
     }
 
     private void copyDTO(ProductDTO dto, Product entity) {
